@@ -1,18 +1,19 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
-import App from './app/App';
+import App from './App';
 import {Provider} from 'react-redux';
-import {store} from './app/store';
+import {store} from './state/store';
 
-ReactDOM.render(
-    <React.StrictMode>
-        <Provider store={store}>
-                <App/>
-        </Provider>
-    </React.StrictMode>,
-    document.getElementById('root'),
+const root = ReactDOM.createRoot(
+    document.getElementById('root') as HTMLElement
+);
+
+root.render(
+    <Provider store={store}>
+    <App />
+    </Provider>
 );
 
 // If you want to start measuring performance in your app, pass a function
