@@ -1,11 +1,7 @@
 module.exports = {
-    "stories": [
-        "../src/**/*.stories.mdx",
-        "../src/**/*.stories.@(js|jsx|ts|tsx)"
-    ],
-    "addons": [
-        "@storybook/addon-links",
-        "@storybook/addon-essentials",
-        "@storybook/preset-create-react-app"
-    ]
-}
+  "stories": ["../src/**/*.stories.mdx", "../src/**/*.stories.@(js|jsx|ts|tsx)"],
+  "addons": ["@storybook/addon-links", "@storybook/addon-essentials", "@storybook/preset-create-react-app"],
+  core: {
+    builder: 'webpack5',
+  },
+};
